// types/interfaces.ts
interface Author {
  _id: string;
  username: string;
  email: string;
}

interface BlogResponse {
  _id: string;
  title: string;
  content: string;
  tags: string[];
  banner: string;
  author: Author;
  updatedAt: string;
  createdAt: string;
}
// ...
interface CreateBlogForm {
  _id?: string;
  title?: string;
  content?: string;
  tags?: string;
  banner?: File | null;
}